requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
generics==3.3.1
idna==2.10
jsonpath==0.82
Markdown==3.3.4
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
